site-visit-report-app/requirements.txt
# Site Visit Report App Dependencies

# Web Framework
flask==3.0.2
flask-cors==4.0.0
fastapi==0.111.0
python-dotenv==1.0.1
gunicorn==21.2.0  # For production deployment
uvicorn==0.29.0   # ASGI server for FastAPI

# Data Processing
pydantic==2.11.0
numpy==1.24.3

# PDF Generation
reportlab==4.0.9
PyPDF2==3.0.1

# AI/ML Components
langchain==0.0.314
langchain-core==0.1.17
langchain-huggingface==0.0.3
faiss-cpu==1.7.4
sentence-transformers==2.5.1

# PDF Processing
pdfplumber==0.10.4

# Security
cryptography==42.0.4
pytest==8.0.0  # For testing

# Monitoring and Performance
prometheus-client==0.19.0
